Separated Mission and MissionComponent classes in 2 files
